requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
pymongo==4.3.3
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.4
pandas==2.0.3
faiss-cpu==1.7.2
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
PyPDF2==3.0.0
